Blackbook_Content_Generator/requirements.txt
streamlit>=1.28.0
cohere>=4.0.0
python-docx>=0.8.11
python-dotenv
pyngrok
